research-tracker/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests>=2.28.0
python-multipart==0.0.6
python-dotenv>=1.0.0
feedparser>=6.0.0
